TODO-resolve swagger exception

+++ Calculator/InputData.cs
@@ -8,14 +8,14 @@ namespace Calculator
 {
     public class InputData
     {
-        public InputData(double value1, double value2, OperationEnum operation)
+        public InputData(int value1, int value2, OperationEnum operation)
         {
             Value1 = value1;
             Value2 = value2;
             Operator = operation;
         }
         public OperationEnum Operator; 
-        public double Value1;
-        public double Value2;
+        public int Value1;
+        public int Value2;
     }
 }
